feat: Added minmaxscaling, chore: moved get_features to utils
